+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 # Feat
 
 - Add blinking info for approval on ledger or trezor [#5331](https://github.com/MyEtherWallet/MyEtherWallet/pull/5331)
+- store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
 
 ### v7.0.2